other driver support + refactoring
