+++ sql/internal/specutil/spec.go
@@ -1,9 +1,11 @@
 package specutil
 
 import (
+	"fmt"
 	"strconv"
 
 	"ariga.io/atlas/schema/schemaspec"
+	"ariga.io/atlas/sql/schema"
 	"ariga.io/atlas/sql/sqlspec"
 )
 
@@ -44,4 +46,65 @@ func ListAttr(k string, litValues ...string) *schemaspec.Attr {
 		K: k,
 		V: lv,
 	}
+}
+
+type doc struct {
+	Tables  []*sqlspec.Table  `spec:"table"`
+	Schemas []*sqlspec.Schema `spec:"schema"`
+}
+
+// Marshal marshals v into an Atlas DDL document using a schemaspec.Marshaler. Marshal uses the given
+// schemaSpec function to convert a *schema.Schema into *sqlspec.Schema and []*sqlspec.Table.
+func Marshal(v interface{}, marshaler schemaspec.Marshaler, schemaSpec func(schem *schema.Schema) (*sqlspec.Schema, []*sqlspec.Table, error)) ([]byte, error) {
+	d := &doc{}
+	switch s := v.(type) {
+	case *schema.Schema:
+		spec, tables, err := schemaSpec(s)
+		if err != nil {
+			return nil, fmt.Errorf("mysql: failed converting schema to spec: %w", err)
+		}
+		d.Tables = tables
+		d.Schemas = []*sqlspec.Schema{spec}
+	case *schema.Realm:
+		for _, s := range s.Schemas {
+			spec, tables, err := schemaSpec(s)
+			if err != nil {
+				return nil, fmt.Errorf("mysql: failed converting schema to spec: %w", err)
+			}
+			d.Tables = append(d.Tables, tables...)
+			d.Schemas = append(d.Schemas, spec)
+		}
+	default:
+		return nil, fmt.Errorf("mysql: failed marshaling spec. %T is not supported", v)
+	}
+	return marshaler.MarshalSpec(d)
+}
+
+// Unmarshal unmarshals an Atlas DDL document using an unmarshaler into v. Unmarshal uses the
+// given convertTable function to convert a *sqlspec.Table into a *schema.Table.
+func Unmarshal(data []byte, unmarshaler schemaspec.Unmarshaler, v interface{}, convertTable func(spec *sqlspec.Table, parent *schema.Schema) (*schema.Table, error)) error {
+	var d doc
+	if err := unmarshaler.UnmarshalSpec(data, &d); err != nil {
+		return err
+	}
+	switch v := v.(type) {
+	case *schema.Realm:
+		realm, err := Realm(d.Schemas, d.Tables, convertTable)
+		if err != nil {
+			return fmt.Errorf("postgres: failed converting to *schema.Realm: %w", err)
+		}
+		*v = *realm
+	case *schema.Schema:
+		if len(d.Schemas) != 1 {
+			return fmt.Errorf("postgres: expecting document to contain a single schema, got %d", len(d.Schemas))
+		}
+		conv, err := Schema(d.Schemas[0], d.Tables, convertTable)
+		if err != nil {
+			return fmt.Errorf("postgres: failed converting to *schema.Schema: %w", err)
+		}
+		*v = *conv
+	default:
+		return fmt.Errorf("postgres: failed unmarshaling spec. %T is not supported", v)
+	}
+	return nil
 }
